fix: move to actions
